Moved all wall classes to new package
