+++ src/main/java/nl/tudelft/model/wall/MovingWall.java
@@ -1,11 +1,13 @@
-package nl.tudelft.model;
+package nl.tudelft.model.wall;
 
 import nl.tudelft.semgroup4.Modifiable;
 
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
 
-public class VerMovingWall extends MovingWall {
+public abstract class MovingWall extends AbstractWall {
+    
+    private int speed;
 
     /**
      * Creates a moving wall, which will move in a vertical
@@ -19,13 +21,21 @@ public class VerMovingWall extends MovingWall {
      * @param speed
      *      - The speed with which the wall moves.
      */
-    public VerMovingWall(Image image, int locX, int locY, int speed) {
-        super(image, locX, locY, speed);
+    public MovingWall(Image image, int locX, int locY, int speed) {
+        super(image, locX, locY);
+        
+        this.speed = speed;
+    }
+    
+    public void setSpeed(int speed) {
+        this.speed = speed;
+    }
+    
+    public int getSpeed() {
+        return speed;
     }
 
     @Override
-    public <T extends Modifiable> void update(T container, int delta) throws SlickException {
-        setLocY(getLocY() + getSpeed());
-    }
+    public  abstract <T extends Modifiable> void update(T container, int delta) throws SlickException;
 
 }
