+++ src/test/java/nl/tudelft/model/WallTest.java
@@ -1,6 +1,8 @@
 package nl.tudelft.model;
 
 import static org.junit.Assert.assertEquals;
+import nl.tudelft.model.wall.AbstractWall;
+import nl.tudelft.model.wall.RegularWall;
 
 import org.junit.Before;
 import org.junit.Test;
